Remaming to Blackboard classic and adding UUID expansion
