--- lms/api_client/blackboard/__init__.py
@@ -1,1 +1,0 @@
-from lms.api_client.blackboard.client import BlackboardClient
