+++ lms/api_client/blackboard_classic/__init__.py
@@ -1,0 +1,1 @@
+from lms.api_client.blackboard_classic.client import BlackboardClassicClient
