+++ lms/api_client/blackboard_classic/client.py
@@ -1,8 +1,8 @@
-from lms.api_client.blackboard.api import APIRoot
+from lms.api_client.blackboard_classic.api import APIRoot
 from lms.api_client.generic_http.oauth2.client import OAuth2Client
 
 
-class BlackboardClient(OAuth2Client):
+class BlackboardClassicClient(OAuth2Client):
     authorization_code_endpoint = "oauth2/authorizationcode"
     access_token_endpoint = "oauth2/token"
